+++ src/main/resources/application.properties
@@ -18,7 +18,7 @@ rules.path: /rules/ArtifactRules-Eiffel-Agen-Version.json
 # these top packages
 logging.level.root: OFF
 logging.level.org.springframework.web: ERROR
-logging.level.com.ericsson.ei: ERROR
+logging.level.com.ericsson.ei: INFO
 
 # Details for connection to RabbitMQ
 rabbitmq.host: localhost
